[ParkingLot] fix: return type of get available position rate should be double

+++ src/main/java/com/parkinglot/ParkingLot.java
@@ -58,7 +58,7 @@ public class ParkingLot {
         return capacity - occupiedPosition;
     }
 
-    public Integer getAvailablePositionRate() {
-        return getRemainingPosition() / capacity;
+    public Double getAvailablePositionRate() {
+        return (double) getRemainingPosition() / capacity;
     }
 }
